minor refactor of path & tests
